backend/requirements.txt
python-dotenv
pymilvus
pypdf

langchain
langchain-openai
langchain-core
langchain-community
langchain-nvidia-ai-endpoints
langchain-milvus
langgraph
langgraph-cli
langgraph-cli[inmem]
langgraph-supervisor

bcrypt==3.2.2
python-multipart
fastapi
yfinance
joblib
scikit-learn==1.6.1
passlib
python-jose
sqlalchemy
psycopg2-binary
pandas
